Ports the code used by Utf8Parser, modified to support things Utf8Parser doesn't but DateTime{Offset{.ParseExact do, such as single-digit offset hours.

Local times (w/o timezone) improve by ~6x.
Times with timezones improve by ~2.5-2.75x.
All cases reduce allocations from 120 bytes across 5 objects down to 0.

Benchmark:
```C#
using System;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Benchmark
{
    private static void Main() => BenchmarkRunner.Run<Benchmark>();

    [Benchmark] public DateTime ParseExactLocal() => DateTime.ParseExact("2018-07-05T22:10:55.5595539", "o", null);
    [Benchmark] public DateTime ParseExactUtc() => DateTime.ParseExact("2018-07-05T22:10:55.5595539Z", "o", null);
    [Benchmark] public DateTime ParseExactOffset() => DateTime.ParseExact("2018-07-05T22:10:55.5595539+12:34", "o", null);
}
```

Before:
```
           Method |     Mean |    Error |   StdDev |   Median |  Gen 0 | Allocated |
----------------- |---------:|---------:|---------:|---------:|-------:|----------:|
  ParseExactLocal | 623.9 ns | 12.42 ns | 28.54 ns | 613.6 ns | 0.0277 |     120 B |
    ParseExactUtc | 962.3 ns | 32.57 ns | 96.03 ns | 932.3 ns | 0.0277 |     120 B |
 ParseExactOffset | 930.7 ns | 18.31 ns | 22.49 ns | 930.7 ns | 0.0277 |     120 B |
```

After:
```
           Method |     Mean |    Error |   StdDev | Allocated |
----------------- |---------:|---------:|---------:|----------:|
  ParseExactLocal | 101.3 ns | 2.205 ns | 3.091 ns |       0 B |
    ParseExactUtc | 349.8 ns | 5.253 ns | 4.656 ns |       0 B |
 ParseExactOffset | 357.7 ns | 3.720 ns | 3.106 ns |       0 B |
```

cc: @jkotas, @danmosemsft, @tarekgh, @joperezr, @joperezr 
Contributes to https://github.com/dotnet/corefx/issues/30612